Corrections:  - ray-sphere intersection  - vec3 cross product and frame build  - lookAt params parsing

+++ project05/src/core/api.h
@@ -58,9 +58,9 @@ class API {
 };
 
 void API::setLookAt(ParamSet & ps) {
-    auto lookAt = ps.findArray<int>(LookAtParams::LOOK_AT);
-    auto lookFrom = ps.findArray<int>(LookAtParams::LOOK_FROM);
-    auto up = ps.findArray<int>(LookAtParams::UP);
+    auto lookAt = ps.findArray<float>(LookAtParams::LOOK_AT);
+    auto lookFrom = ps.findArray<float>(LookAtParams::LOOK_FROM);
+    auto up = ps.findArray<float>(LookAtParams::UP);
 
     ro.lookAt = LookAt(Point3D(lookFrom[0], lookFrom[1], lookFrom[2]),
                        Point3D(lookAt[0], lookAt[1], lookAt[2]),
